unloading board class by factoring board initialization into a new class
